genericapiserver: move pkg/auth/handlers into filters
