+++ pkg/BUILD
@@ -31,7 +31,6 @@ filegroup(
         "//pkg/apis/rbac:all-srcs",
         "//pkg/apis/storage:all-srcs",
         "//pkg/auth/authorizer/abac:all-srcs",
-        "//pkg/auth/handlers:all-srcs",
         "//pkg/auth/user:all-srcs",
         "//pkg/capabilities:all-srcs",
         "//pkg/client/cache:all-srcs",
